Alphabetize members in member partial
Resolves #186

To do this, I added the has_many :members association to both Group and
Meeting and changed the notifications/_members partial so that it
can receive an instance of either group or meeting as the local variable
called group. Using this instance, the partial can get anything it needs
since both group and meeting respond to #leaders and #members. Instead
of passing a second :is_meeting_member local, we can check whether group
is an instance of Group.

In order to implement the actual alphabetization part, I added the
default ordering on name to the members association to both group and
meeting models.