Extracting GroupNotifier service object
This change removes duplicated code in the join, update and create
methods of the groups_controller by extracting a service object to
notify the relevant members of each group.

Testing the GroupNotifier was a little messy because of how much
interaction there was with collaborators, so I ended up using a lot of
stubs.