Adding more robust setup script from gist
https://gist.github.com/gvarela/9370838

Note this is a WIP. I'm just going to commit this as-is and then attempt
to customize it for our needs. I thought it would be helpful to separate
these commits so that we will have a diff showing what is from the gist
and what I changed. Calling this setup_new for now. We can use it to
replace bin/setup once we are satisfied with it.