Updating database configuration
* Removing verify_database_configuration - This function copied over the
database configuration, which we already have in
copy_example_config_files. It also runs rake db:create:all, which is
handled in rake db:setup (which we run later), so we don't need this
here. I added the error handling from this to the rake db:setup call and
improved the error message.

* Checking for existence of config files before copying examples over