Disable cops that relate to code smell
For all its strengths, Rubocop isn't a great tool for checking for code smell.
When a developer is given an error like "Avoid using update_all because it skips
validations" in a situation where they have made a conscious and informed
decision, they are left in the unfortunate position of choosing between ignoring
the error and littering the code with "rubocop:disable" comments.

If code smell detection is desired, it would be better to implement a tool
designed for this function (e.g. reek) and then enforcing that Rubocop is the
law whereas code smell advice can be taken or ignored.