Stop method execution if comment does not save
If the comment has not saved, the respond_to method is called, however, no
return statement was given so the rest of the comment method continues to
be processed.

This would also be why it was necessary to wrap the final call to respond_to
in a `if @comment.save`, despite the fact that @comment has not been modified
and therefore has nothing to be saved.