Update README (#598)
* Nokogiri is installed by bundler

README tells people to install nokogiri using `gem install` if it isn't already
installed. Nokogiri is, however, installed by bundler and installing with a
simple `gem install` command could result in installing a version different from
that used in the project and the user will end up with two copies installed.

* Remove README direction to set local ruby version

Calling `rbenv local [version]` adds the version number to a .ruby-version file.
Since there is a already .ruby-version file in the project, this step is
unnecessary and just another bit of documentation to keep up-to-date

* Remove README direction to update Rails

Both the installation and version maintenance of Ruby on Rails is managed by
bundler so there should be no reason for people to manually update their version

* Update PostgreSQL on Linux installation instructions

* Install PostgreSQL before Ruby

You can't completely the Ruby on Rails installation if you do not have
PostgreSQL installed. The pg gem has C bindings which rely on the installed
PostgreSQL instance.

* Add instructions for adding the current user to PostgreSQL