[wip][skip ci][iss #606]info on group page
* display more detailed info about groups by using text from FAQ
* wip

I changed part of the EN text (pending approval) to:
You aren't a part of any groups yet, but you
can click %{icon} to create one!

Formerly, it was:
You aren't a part of any groups yet! Click to
create one or join any available groups below.

Let me know if that's ok, @julianguyen, and I'll get the translations
and test the patch.